requirements.txt
asgiref==3.3.1
astroid==2.2.5
boto==2.49.0
boto3==1.12.21
botocore==1.15.21
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.1
DAWG-Python==0.7.2
dj-database-url==0.5.0
Django==3.0.8
django-appconf==1.0.4
django-compressor==2.4
django-cors-headers==3.5.0
djangorestframework==3.10.1
docopt==0.6.2
docutils==0.15.2
gensim==3.8.1
idna==2.9
inflect==5.0.2
isort==4.3.21
gunicorn==20.0.4
jmespath==0.9.5
joblib==0.14.1
lazy-object-proxy==1.4.1
lxml==4.5.0
mccabe==0.6.1
nltk==3.5
nodeenv==1.5.0
num2words==0.5.10
numpy==1.18.1
pdfminer.six==20200124
Pillow==7.0.0
psycopg2-binary==2.8.6
pycryptodome==3.9.7
pylint==2.3.1
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-docx==0.8.10
pytz==2019.1
rcssmin==1.0.6
regex==2020.4.4
requests==2.23.0
rjsmin==1.1.0
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.12.0
sklearn==0.0
smart-open==1.9.0
sorl-thumbnail==12.7.0
sortedcontainers==2.1.0
sqlparse==0.3.0
stripe==2.55.1
tqdm==4.45.0
typed-ast==1.4.0
urllib3==1.25.8
whitenoise==5.2.0
wrapt==1.11.2
